+++ src/00_raster.cpp
@@ -6,17 +6,17 @@ using namespace Rcpp;
 
 // [[Rcpp::export]]
 List raster_gcp_gdal_cpp(CharacterVector dsn) {
-  return gdallibrary::gdal_raster_gcp(dsn);
+  return gdalraster::gdal_raster_gcp(dsn);
 }
 
 // [[Rcpp::export]]
 List raster_info_gdal_cpp(CharacterVector dsn, LogicalVector min_max) {
-  return gdallibrary::gdal_raster_info(dsn, min_max);
+  return gdalraster::gdal_raster_info(dsn, min_max);
 }
 
 // [[Rcpp::export]]
 NumericVector raster_extent_cpp(CharacterVector dsn) {
-  return gdallibrary::gdal_extent_only(dsn); 
+  return gdalraster::gdal_extent_only(dsn); 
 }
 
 // [[Rcpp::export]]
@@ -25,19 +25,19 @@ List raster_io_gdal_cpp(CharacterVector dsn,
                         IntegerVector band,
                         CharacterVector resample,
                         CharacterVector band_output_type) {
-  return gdallibrary::gdal_raster_io(dsn, window, band, resample, band_output_type);
+  return gdalraster::gdal_raster_io(dsn, window, band, resample, band_output_type);
 }
 
 // [[Rcpp::export]]
 CharacterVector sds_list_gdal_cpp(CharacterVector dsn) {
-  return gdallibrary::gdal_sds_list(dsn[0]);
+  return gdalraster::gdal_sds_list(dsn[0]);
 }
 
 // [[Rcpp::export]]
 List sds_list_list_gdal_cpp(CharacterVector dsn) {
   List outlist(dsn.length()); 
   for (int i = 0; i < dsn.length(); i++) {
-    outlist[i] = gdallibrary::gdal_sds_list(dsn[i]);
+    outlist[i] = gdalraster::gdal_sds_list(dsn[i]);
   }
   return outlist; 
 }